+++ vscode-react/src/app.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from "react";
-import { AKHeader } from "@react-components";
+import { Header } from "@react-components";
 import { AppStateProvider } from "@react-context";
-import { AKDeployments, AKSessions } from "@react-sections";
+import { Deployments, Sessions } from "@react-sections";
 import "./app.css";
 import SplitPane from "split-pane-react";
 import "split-pane-react/esm/themes/default.css";
@@ -21,7 +21,7 @@ function App() {
 		<AppStateProvider>
 			<main ref={ref}>
 				<div className="flex flex-col w-full">
-					<AKHeader />
+					<Header />
 					<div className="h-[calc(100vh-6vh)]">
 						<SplitPane
 							split="horizontal"
@@ -30,10 +30,10 @@ function App() {
 							sashRender={() => <hr className="bg-vscode-editor-background h-3" />}
 						>
 							<div>
-								<AKDeployments height={sizes[0]} />
+								<Deployments height={sizes[0]} />
 							</div>
 							<div>
-								<AKSessions height={sizes[1]} />
+								<Sessions height={sizes[1]} />
 							</div>
 						</SplitPane>
 					</div>
