--- vscode-react/src/components/AKTable/aKTableMessage.component.tsx
@@ -1,9 +1,0 @@
-import React, { ReactNode } from "react";
-
-interface AKTableMessageProps {
-	children: ReactNode;
-}
-
-export const AKTableMessage = ({ children }: AKTableMessageProps) => {
-	return <div className="w-full text-center m-4">{children}</div>;
-};
