--- vscode-react/src/components/AKTable/index.ts
@@ -1,6 +1,0 @@
-export { AKTable } from "./aKTable.component";
-export { AKTableCell } from "./aKTableCell.component";
-export { AKTableHeader } from "./aKTableHeader.component";
-export { AKTableRow } from "./aKTableRow.component";
-export { AKTableHeaderCell } from "./aKTableHeaderCell.component";
-export { AKTableMessage } from "./aKTableMessage.component";
