--- vscode-react/src/components/aKOverlay.component.tsx
@@ -1,11 +1,0 @@
-type AKOverlayProps = {
-	onOverlayClick?: () => void;
-	isVisibile: boolean;
-};
-export const AKOverlay = ({ onOverlayClick = () => {}, isVisibile }: AKOverlayProps) =>
-	isVisibile && (
-		<div
-			className="absolute h-screen w-screen top-0 left-0 bg-black opacity-50 z-50"
-			onClick={() => onOverlayClick()}
-		></div>
-	);
