+++ vscode-react/src/components/atoms/table/tableCell.component.tsx
@@ -1,0 +1,18 @@
+import React, { ReactNode } from "react";
+import clsx from "clsx";
+
+interface CellProps {
+	children: ReactNode;
+	onClick?: () => void;
+	classes?: string[];
+}
+
+export const TableCell = ({ children, onClick, classes }: CellProps) => {
+	const baseClass = "text-center p-1";
+	const cellClass = clsx(baseClass, classes);
+	return (
+		<td className={cellClass} onClick={onClick}>
+			{children}
+		</td>
+	);
+};
