+++ vscode-react/src/components/atoms/table/tableMessage.component.tsx
@@ -1,0 +1,9 @@
+import React, { ReactNode } from "react";
+
+interface TableMessageProps {
+	children: ReactNode;
+}
+
+export const TableMessage = ({ children }: TableMessageProps) => {
+	return <div className="w-full text-center m-4">{children}</div>;
+};
