--- vscode-react/src/components/index.ts
@@ -1,16 +1,0 @@
-export { AKButton } from "@react-components/aKButton.component";
-export { AKLogo } from "@react-components/aKLogo.component";
-export { AKDeploymentState } from "@react-components/aKDeploymentState.component";
-export { AKSessionState } from "@react-components/aKSessiontState.component";
-export { AKModal } from "@react-components/aKModal.component";
-export { AKDeploymentTableHeader } from "@react-components/aKDeploymentTableHeader.component";
-export { AKDeploymentTableBody } from "@react-components/aKDeploymentTableBody.component";
-export { AKMonacoEditorModal } from "@react-components/akMonacoEditorModal.component";
-export { ExecutePopper } from "@react-components/aKDeploymentsExecutePopper.component";
-export { DeletePopper } from "@react-components/aKDeletePopper.component";
-export { PopperComponent } from "@react-components/aKPopper.component";
-export { ProjectSettingsPopper } from "@react-components/aKProjectSettingsPopper.component";
-export { AKOverlay } from "@react-components/aKOverlay.component";
-export { AKHeader } from "@react-components/aKHeader.component";
-export { AKSessionActions } from "@react-components/aKSessionActions.component";
-export { AKSessionsTableRow } from "@react-components/aKSessionTableRow.component";
