+++ vscode-react/src/components/organisms/aKTable.component.tsx
@@ -1,0 +1,14 @@
+import React, { ReactNode } from "react";
+import clsx from "clsx";
+
+interface TableProps {
+	children: ReactNode;
+	classes?: string;
+}
+
+export const Table = ({ children, classes }: TableProps) => {
+	const baseClass = ["w-full"];
+	const tableClass = clsx(baseClass, [classes]);
+
+	return <table className={tableClass}>{children}</table>;
+};
