+++ vscode-react/src/components/sessions/pages/aKSessions.section.tsx
@@ -3,10 +3,10 @@ import { MessageType, SessionStateType } from "@enums";
 import { translate } from "@i18n";
 import { SessionSectionViewModel } from "@models/views";
 import RotateIcon from "@react-assets/icons/rotate.svg?react";
-import { SessionsTableBody } from "@react-components/aKSessionTableBody.component";
+import { SessionsTableBody } from "@react-components/sessions/organisms/aKSessionTableBody.component";
 import { TableMessage } from "@react-components/Table";
 import { useAppState } from "@react-context";
-import { useIncomingMessageHandler, useForceRerender } from "@react-hooks";
+import { useIncomingMessageHandler, useForceRerender, useDelayedLoading } from "@react-hooks";
 import { sendMessage } from "@react-utilities";
 
 export const Sessions = ({ height }: { height: string | number }) => {
@@ -14,7 +14,8 @@ export const Sessions = ({ height }: { height: string | number }) => {
 	const [selectedSession, setSelectedSession] = useState<string | undefined>("");
 	const [stateFilter, setStateFilter] = useState<string>();
 	const [liveTailState, setLiveTailState] = useState<boolean>(true);
-	const [{ delayedLoading }] = useAppState();
+	const [{ loading }] = useAppState();
+	const delayedLoading = useDelayedLoading(loading, 1000);
 
 	const { sessions, showLiveTail, lastDeployment } = sessionsSection || {};
