+++ vscode-react/src/sections/index.ts
@@ -1,2 +1,2 @@
-export { AKDeployments } from "@react-sections/aKDeployments.section";
-export { AKSessions } from "@react-sections/aKSessions.section";
+export { Deployments } from "@react-sections/aKDeployments.section";
+export { Sessions } from "@react-sections/aKSessions.section";
